Apply the patch for DIRMINA-732
git-svn-id: https://svn.apache.org/repos/asf/mina/trunk@899366 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mina/transport/socket/nio/NioDatagramAcceptor.java
@@ -174,7 +174,7 @@ public final class NioDatagramAcceptor
     }
 
     @Override
-    protected int select(int timeout) throws Exception {
+    protected int select(long timeout) throws Exception {
         return selector.select(timeout);
     }
